Use ToolCallUpdate instead of a ToolCall for RequestPermissions
In practice it is common for agents to create the tool call before they
request permissions. It would be nice if requesting permission only took
an update so that we don't have to copy over the entire tool call and
can instead.

This should be backwards compatible on the protocol side, since they
would still be sending all fields.

However, we may need to make sure on the client side that we either a)
require that a tool call is pushed before an update or b) handle missing
fields gracefully with some form of default.

The latter may be a reason for not doing this and just dealing with the
fact that we have to re-create the entire tool call.